requirements.txt
pwntools>=4.11.0
